chore: pull out network switching functions into a hook
